refactor: switch MCP to direct self-reexport imports
