+++ packages/opencode/src/command/command.ts
@@ -6,7 +6,7 @@ import { SessionID, MessageID } from "@/session/schema"
 import { Effect, Layer, Context } from "effect"
 import z from "zod"
 import { Config } from "../config"
-import { MCP } from "../mcp"
+import { MCP } from "../mcp/mcp"
 import { Skill } from "../skill"
 import PROMPT_INITIALIZE from "./template/initialize.txt"
 import PROMPT_REVIEW from "./template/review.txt"
