+++ packages/opencode/src/effect/app-runtime.ts
@@ -34,7 +34,7 @@ import { SessionPrompt } from "@/session/prompt"
 import { Instruction } from "@/session/instruction"
 import { LLM } from "@/session/llm"
 import { LSP } from "@/lsp"
-import { MCP } from "@/mcp"
+import { MCP } from "@/mcp/mcp"
 import { McpAuth } from "@/mcp/auth"
 import { Command } from "@/command"
 import { Truncate } from "@/tool"
