--- packages/opencode/src/mcp/index.ts
@@ -1,1 +1,0 @@
-export * as MCP from "./mcp"
