+++ packages/opencode/src/server/instance/experimental.ts
@@ -6,7 +6,7 @@ import { ToolRegistry } from "../../tool"
 import { Worktree } from "../../worktree"
 import { Instance } from "../../project/instance"
 import { Project } from "../../project"
-import { MCP } from "../../mcp"
+import { MCP } from "../../mcp/mcp"
 import { Session } from "../../session"
 import { Config } from "../../config"
 import { ConsoleState } from "../../config/console-state"
