+++ packages/opencode/src/server/instance/mcp.ts
@@ -1,7 +1,7 @@
 import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
-import { MCP } from "../../mcp"
+import { MCP } from "../../mcp/mcp"
 import { Config } from "../../config"
 import { AppRuntime } from "../../effect/app-runtime"
 import { errors } from "../error"
