+++ packages/opencode/src/session/prompt.ts
@@ -20,7 +20,7 @@ import PROMPT_PLAN from "../session/prompt/plan.txt"
 import BUILD_SWITCH from "../session/prompt/build-switch.txt"
 import MAX_STEPS from "../session/prompt/max-steps.txt"
 import { ToolRegistry } from "../tool"
-import { MCP } from "../mcp"
+import { MCP } from "../mcp/mcp"
 import { LSP } from "../lsp"
 import { FileTime } from "../file/time"
 import { Flag } from "../flag/flag"
