+++ packages/opencode/test/session/prompt-effect.test.ts
@@ -9,7 +9,7 @@ import { Command } from "../../src/command"
 import { Config } from "../../src/config"
 import { FileTime } from "../../src/file/time"
 import { LSP } from "../../src/lsp"
-import { MCP } from "../../src/mcp"
+import { MCP } from "../../src/mcp/mcp"
 import { Permission } from "../../src/permission"
 import { Plugin } from "../../src/plugin"
 import { Provider as ProviderSvc } from "../../src/provider"
